+++ src/main/java/com/esliceu/movies/controllers/ProductionCompanyController.java
@@ -1,7 +1,10 @@
 package com.esliceu.movies.controllers;
 
+import com.esliceu.movies.models.Permission;
 import com.esliceu.movies.models.Production_Company;
+import com.esliceu.movies.services.PermissionService;
 import com.esliceu.movies.services.ProductionCompanyService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,6 +15,9 @@ public class ProductionCompanyController {
 
     @Autowired
     private ProductionCompanyService productionCompanyService;
+
+    @Autowired
+    PermissionService permissionService;
 
     @GetMapping("/productionCompany")
     public String listProductionCompanies(Model model) {
@@ -44,8 +50,10 @@ public class ProductionCompanyController {
     }
 
     @GetMapping("/deleteProductionCompany/{id}")
-    public String deleteProductionCompany(@PathVariable int id) {
-        productionCompanyService.deleteById(id);
+    public String deleteProductionCompany(@PathVariable int id, HttpSession session) {
+        int userId = (int) session.getAttribute("loggedInUserId");
+        boolean canDelete = permissionService.isUserAuthorized(userId, Permission.permission_name.remove_movie);
+        if (canDelete) productionCompanyService.deleteById(id);
         return "redirect:/productionCompany";
     }
 }
